+++ database/migrations/2023_07_24_061923_create_orders_table.php
@@ -18,7 +18,7 @@ class CreateOrdersTable extends Migration
             $table->unsignedBigInteger('customer_id');
             $table->foreign('customer_id')->references('id')->on('customers')->onDelete('cascade');
             $table->json('items');
-            $table->float('total');
+            $table->decimal('total');
             $table->timestamps();
         });
     }
